rename simulate drilldown.
