+++ packages/client/hmi-client/src/components/workflow/ops/simulate-ciemss/mod.ts
@@ -1,6 +1,6 @@
 import { SimulateCiemssOperation as operation } from './simulate-ciemss-operation';
 import node from './tera-simulate-node-ciemss.vue';
-import drilldown from './tera-simulate-ciemss.vue';
+import drilldown from './tera-simulate-ciemss-drilldown.vue';
 
 const name = operation.name;
